+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 - BC-195 - fixed course filter for student task dashboard
 - BC-94 - clean up the colorsystem and fix some UI issues
 - BC-313 - fixed active tab mobile behaviour and color
+- BC-322 - fixed race condition while loading school in school settings page
 
 ## [26.10.1] - 2021-09-07